requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
faiss-cpu
pypdf
tiktoken
groq
loguru
scikit-learn
numpy
